Refactored my_solutions directory structure
