+++ my_solutions/src/package98e3/OneFourTwoOne.java
@@ -1,4 +1,4 @@
-package ic.doc;
+package package98e3;
 
 public class OneFourTwoOne {
