--- my_solutions/template/test/ic/doc/templateTest.java
@@ -1,12 +1,0 @@
-package ic.doc;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class templateTest {
-
-    public void firstTest() {
-
-    }
-}
