+++ solutions/a22c.md
@@ -1,6 +1,6 @@
 [Back to questions](../README.md)
 
-## Solution to [a22c](../questions/a22c): *No duplicate email addresses*
+## Solution to [a22c](../questions/a22c.md): *No duplicate email addresses*
 
 See code at `solutions/code/tutorialquestions/questiona22c`
 
@@ -28,7 +28,7 @@ public EmailAddress(String identifier) throws DuplicateEmailAddressException;
 
 Because the constructors of `IndividualEmailAddress` and `GroupEmailAddress` both call the constructor of `EmailAddress`, these constructors are also declared as throwing `DuplicateEmailAddressException`.
 
-`Demo` shows this exception being thrown in practice: the identifier `wayne@doc.ic.ac.uk` is used in the creation of two different
+`Demo` shows this exception being thrown in practice: the identifier `wayne@package98e3.ic.ac.uk` is used in the creation of two different
 email addresses.
 
 **Advanced:** The package `checkforcycles` includes another exception class:
